src: move impact engine inside.
